Fold Analytics into the hiring hub; cards carry real work items
The hub is now the cycle overview: the pipeline section (status pie,
cycle/domain selectors, application drill-down — formerly /hiring/analytics)
renders inline for Core/DomainLead, behind the same confidentiality gate.
The Analytics pill is gone and /hiring/analytics redirects to /hiring with
its params intact; the pie/selector controls navigate relative to the
current path.

Personal cards stop mirroring the pill row: My Reviews lists the actual
pending reviews (applicant + domain, deep-linked) with a +N-more overflow,
My Interviews rows gain the applicant name, the waitlist count folds into
Cycle health, and the health card loses its button since the pipeline sits
directly below.
